Move parse and resolve up a level
