+++ dhall/src/semantics/mod.rs
@@ -1,6 +1,7 @@
 pub mod builtins;
 pub mod nze;
-pub mod phase;
+pub mod parse;
+pub mod resolve;
 pub mod tck;
 pub(crate) use self::builtins::*;
 pub(crate) use self::nze::*;
