--- dhall/src/semantics/phase/mod.rs
@@ -1,2 +1,0 @@
-pub(crate) mod parse;
-pub(crate) mod resolve;
